pwhash: use uint64_t for opslimit parameter

+++ pwhash.go
@@ -47,7 +47,7 @@ func PwHashDeterministic(h_len int, passwd string, ctx string, master_key []byte
 		C.size_t(len(passwd)),
 		(*C.char)(unsafe.Pointer(&cCtx[0])),
 		(*C.uint8_t)(&master_key[0]),
-		C.size_t(opslimit),
+		C.uint64_t(opslimit),
 		C.size_t(PwHashDeterministicMemLimit),
 		C.uint8_t(PwHashDeterministicThreads)))
